+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # Dominion 0.6.6
 
-Handlanger funktioniert gar nicht
+- Handlanger funktioniert gar nicht
+- Höflinge geht gar nicht
+- Adelige +2 Aktionen geht nicht; 
+- Keller/Adelige nachziehen wird beim Geld nicht berücksichtigt
+- Baron gibt kein Geld
+Mühle gibt zu wenig Geld
 Bürokrat mit Rückmeldung
 dominion.css update
-Mühle gibt zu wenig Geld
-Baron gibt kein Geld
 Aktion kaufen: Karten umgekehrt anzeigen
 Karten zurücklegen stimmt nicht
 Start: "Spieler sucht Kartenset aus" fehlt
+✓ tut gerade Nobles
 
 
 # Dominion 0.6.5